Add files via upload
This was my original idea for this project. Pictured is an R/C car designed to be able to jump as well a roll with the likeness of a frog. It did not pass to the fabrication stage so it remained only a concept design as my group decided on a penguin for our final concept.